+++ src/JsonCodec.php
@@ -34,6 +34,26 @@ class JsonCodec implements JsonCodecInterface {
 	protected const TYPE_ANNOTATION = '_type_';
 
 	/**
+	 * Prefix used to distinguish abbreviations from class names.
+	 */
+	protected const ABBREV_PREFIX = '@';
+
+	/**
+	 * Maps abbreviation names to hint abbreviations.  Keys are prefixed
+	 * with self::ABBREV_PREFIX for faster lookup.
+	 * @see ::addAbbrev()
+	 * @var array<string,Abbrev>
+	 */
+	protected array $abbrevToHintMap = [];
+
+	/**
+	 * Maps PHP class names to abbreviations
+	 * @see ::addAbbrev()
+	 * @var array<class-string,Abbrev>
+	 */
+	protected array $classToAbbrevMap = [];
+
+	/**
 	 * @param ?ContainerInterface $serviceContainer
 	 */
 	public function __construct( ?ContainerInterface $serviceContainer = null ) {
@@ -70,7 +90,7 @@ class JsonCodec implements JsonCodecInterface {
 	 * guide deserialization.
 	 *
 	 * @param mixed|null $value
-	 * @param class-string|Hint|null $classHint An optional hint to
+	 * @param class-string|Hint|Abbrev|null $classHint An optional hint to
 	 *   the type of the encoded object.  If this is provided and matches
 	 *   the type of $value, then explicit type information will be omitted
 	 *   from the generated JSON, which saves some space.
@@ -94,7 +114,7 @@ class JsonCodec implements JsonCodecInterface {
 	 * objects serialized with explicit classes.
 	 *
 	 * @param string $json A JSON-encoded string
-	 * @param class-string|Hint|null $classHint An optional hint to
+	 * @param class-string|Hint|Abbrev|null $classHint An optional hint to
 	 *   the type of the encoded object.  In the absence of explicit
 	 *   type information in the JSON, this will be used as the type of
 	 *   the created object.
@@ -162,6 +182,64 @@ class JsonCodec implements JsonCodecInterface {
 	}
 
 	/**
+	 * This supports cross-platform schemas by decoupling a hint from
+	 * the actual PHP class name.  If the `_type_` specified in the
+	 * JSON starts with `self::ABBREV_PREFIX` it is looked up in the
+	 * abbreviation map to obtain a hint.
+	 *
+	 * Only abbreviations that map to `class-string` are used for encoding.
+	 * If you wish to add an `class-string` abbreviation that is only for
+	 * decode, then building a simple Hint with ONLY_FOR_DECODE is sufficient:
+	 * ```
+	 * $codec->addAbbrev($name, Hint::build($className, HintType::ONLY_FOR_DECODE));
+	 * ```
+	 * This allows for forward-compatibility with a future encoding which
+	 * uses abbreviations (as usual for the ONLY_FOR_DECODE hint), but it
+	 * can also be used to allow multiple abbreviations for the same class
+	 * name, as long at most one of them is registered without
+	 * ONLY_FOR_DECODE.
+	 * @param string $name The abbreviation name (unprefixed)
+	 * @param class-string|Hint $hint The PHP class name or hint
+	 *  to be abbreviated
+	 */
+	public function addAbbrev( string $name, string|Hint $hint ): Abbrev {
+		$abbrev = new Abbrev( $name, $hint );
+		// Each abbreviation must map to a single class name for decode,
+		// although we can have a class map to multiple abbreviations for
+		// encode (only the last registered is used).
+		$key = self::ABBREV_PREFIX . $name;
+		$existing = $this->abbrevToHintMap[$key] ?? null;
+		if (
+			$existing !== null &&
+			!$abbrev->isSameAs( $existing )
+		) {
+			throw new InvalidArgumentException(
+				"conflicting abbreviation for {$name}: {$existing->hint} != {$abbrev->hint}"
+			);
+		}
+		$this->abbrevToHintMap[$key] = $abbrev;
+		// Only abbreviations for bare class names (not hints) are serialized.
+		if ( is_string( $hint ) ) {
+			$existing = $this->classToAbbrevMap[$hint] ?? null;
+			if ( $existing !== null && !$abbrev->isSameAs( $existing ) ) {
+				throw new InvalidArgumentException(
+					"too many abbreviations for {$hint}: {$existing->name}, {$abbrev->name}"
+				);
+			}
+			$this->classToAbbrevMap[$hint] = $abbrev;
+		}
+		return $abbrev;
+	}
+
+	/**
+	 * Return an abbreviation registered for the given abbrevation name.
+	 */
+	public function getAbbrev( string $name ): ?Abbrev {
+		$key = self::ABBREV_PREFIX . $name;
+		return $this->abbrevToHintMap[$key] ?? null;
+	}
+
+	/**
 	 * Recursively converts a given object to an associative array
 	 * which can be json-encoded.  (When embedding an object into
 	 * another context it is sometimes useful to have the array
@@ -177,7 +255,7 @@ class JsonCodec implements JsonCodecInterface {
 	 * guide deserialization.
 	 *
 	 * @param mixed|null $value
-	 * @param class-string|Hint|null $classHint An optional hint to
+	 * @param class-string|Hint|Abbrev|null $classHint An optional hint to
 	 *   the type of the encoded object.  If this is provided and matches
 	 *   the type of $value, then explicit type information will be omitted
 	 *   from the generated JSON, which saves some space.
@@ -192,6 +270,9 @@ class JsonCodec implements JsonCodecInterface {
 		$arrayClassHint = null;
 		$forceBraces = null;
 		$allowInherited = false;
+		if ( $classHint instanceof Abbrev ) {
+			$classHint = $classHint->hint;
+		}
 		if ( is_string( $classHint ) && str_ends_with( $classHint, '[]' ) ) {
 			// back-compat
 			$classHint = new Hint( substr( $classHint, 0, -2 ), HintType::LIST );
@@ -318,7 +399,7 @@ class JsonCodec implements JsonCodecInterface {
 	 * objects serialized with explicit classes.
 	 *
 	 * @param mixed|null $json
-	 * @param class-string|Hint|null $classHint An optional hint to
+	 * @param class-string|Hint|Abbrev|null $classHint An optional hint to
 	 *   the type of the encoded object.  In the absence of explicit
 	 *   type information in the JSON, this will be used as the type of
 	 *   the created object.
@@ -332,6 +413,9 @@ class JsonCodec implements JsonCodecInterface {
 
 		// Process class hints
 		$arrayClassHint = null;
+		if ( $classHint instanceof Abbrev ) {
+			$classHint = $classHint->hint;
+		}
 		while ( $classHint instanceof Hint ) {
 			if ( $classHint->modifier === HintType::LIST ) {
 				$arrayClassHint = $classHint->parent;
@@ -458,7 +542,10 @@ class JsonCodec implements JsonCodecInterface {
 		// we need to escape it (by hoisting into a child array).
 		if ( array_key_exists( self::TYPE_ANNOTATION, $value ) ) {
 			if ( !self::classEquals( $className, $classHint ) ) {
-				$value[self::TYPE_ANNOTATION] = [ $value[self::TYPE_ANNOTATION], $className ];
+				$abbrev = $this->classToAbbrevMap[$className] ?? null;
+				$type = ( $abbrev === null ) ? $className :
+					  ( self::ABBREV_PREFIX . $abbrev->name );
+				$value[self::TYPE_ANNOTATION] = [ $value[self::TYPE_ANNOTATION], $type ];
 			} else {
 				// Omit $className since it matches the $classHint, but we still
 				// need to escape the field to make it clear it was marked.
@@ -471,7 +558,10 @@ class JsonCodec implements JsonCodecInterface {
 			!self::classEquals( $className, $classHint )
 		) {
 			// Include the type annotation if it doesn't match the hint
-			$value[self::TYPE_ANNOTATION] = $className;
+			$abbrev = $this->classToAbbrevMap[$className] ?? null;
+			$type = ( $abbrev === null ) ? $className :
+				  ( self::ABBREV_PREFIX . $abbrev->name );
+			$value[self::TYPE_ANNOTATION] = $type;
 		}
 	}
 
@@ -493,14 +583,33 @@ class JsonCodec implements JsonCodecInterface {
 	 *  'array' if the value was a "complex" or "false mark" array.
 	 */
 	protected function unmarkArray( array &$value, ?string $classHint ): string {
-		$className = $value[self::TYPE_ANNOTATION] ?? $classHint;
+		$abbrevOrClass = $value[self::TYPE_ANNOTATION] ?? null;
 		// Remove our marker and restore the previous state of the
 		// json array (restoring a pre-existing field if needed)
-		if ( is_array( $className ) ) {
-			$value[self::TYPE_ANNOTATION] = $className[0];
-			$className = $className[1] ?? $classHint;
+		if ( is_array( $abbrevOrClass ) ) {
+			[ $oldValue, $abbrevOrClass ] = array_pad( $abbrevOrClass, 2, null );
+			$value[self::TYPE_ANNOTATION] = $oldValue;
 		} else {
 			unset( $value[self::TYPE_ANNOTATION] );
+		}
+		if ( str_starts_with( $abbrevOrClass ?? '', self::ABBREV_PREFIX ) ) {
+			$abbrev = $this->abbrevToHintMap[$abbrevOrClass] ?? null;
+			if ( $abbrev === null ) {
+				throw new InvalidArgumentException(
+					"Unknown abbreviation: $abbrevOrClass"
+				);
+			}
+			$className = $abbrev->hint;
+			while ( $className instanceof Hint && $className->modifier === HintType::ONLY_FOR_DECODE ) {
+				$className = $className->parent;
+			}
+			if ( !is_string( $className ) ) {
+				throw new InvalidArgumentException(
+					"Abbreviation used to encode a hint, not a class"
+				);
+			}
+		} else {
+			$className = $abbrevOrClass ?? $classHint;
 		}
 		return $className;
 	}